Reorganize source files into packages
