+++ app/src/main/AndroidManifest.xml
@@ -50,7 +50,7 @@
             android:grantUriPermissions="true"></provider>
 
         <activity
-            android:name=".SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
