--- app/src/main/kotlin/com/patrykmis/bar/AudioFormatExtension.kt
@@ -1,6 +1,0 @@
-package com.patrykmis.bar
-
-import android.media.AudioFormat
-
-val AudioFormat.frameSizeInBytesCompat: Int
-    get() = frameSizeInBytes
