+++ app/src/main/kotlin/com/patrykmis/bar/RecorderMicTileService.kt
@@ -7,6 +7,7 @@ import android.os.Build
 import android.service.quicksettings.Tile
 import android.service.quicksettings.TileService
 import android.util.Log
+import com.patrykmis.bar.settings.SettingsActivity
 
 class RecorderMicTileService : TileService() {
     companion object {
