+++ app/src/main/kotlin/com/patrykmis/bar/extension/DocumentFileExtensions.kt
@@ -1,4 +1,4 @@
-package com.patrykmis.bar
+package com.patrykmis.bar.extension
 
 import android.content.ContentResolver
 import android.content.Context
