+++ app/src/main/kotlin/com/patrykmis/bar/output/OutputDirUtils.kt
@@ -1,4 +1,4 @@
-package com.patrykmis.bar
+package com.patrykmis.bar.output
 
 import android.content.Context
 import android.net.Uri
@@ -8,6 +8,7 @@ import android.system.Os
 import android.system.OsConstants
 import android.util.Log
 import androidx.documentfile.provider.DocumentFile
+import com.patrykmis.bar.Preferences
 import java.io.IOException
 
 class OutputDirUtils(private val context: Context, private val redactor: Redactor) {
