+++ app/src/main/kotlin/com/patrykmis/bar/settings/SettingsActivity.kt
@@ -1,0 +1,22 @@
+package com.patrykmis.bar.settings
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.patrykmis.bar.R
+
+class SettingsActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.settings_activity)
+        if (savedInstanceState == null) {
+            supportFragmentManager
+                .beginTransaction()
+                .replace(R.id.settings, SettingsFragment())
+                .commit()
+        }
+
+        setSupportActionBar(findViewById(R.id.toolbar))
+
+        setTitle(R.string.app_name_full)
+    }
+}
